+++ CHANGELOG.md
@@ -68,6 +68,8 @@ and this project adheres to
   consistency with other types like `Response`.
 - cosmwasm-vm: `Instance::required_features` changed from a property to a getter
   method.
+- cosmwasm-vm: Add `required_features` field to `AnalysisReport` which is
+  returned by `Cache::analyze`.
 
 [#995]: https://github.com/CosmWasm/cosmwasm/pull/995